Trait mp_snos_output::SnosCodec

pub trait SnosCodec: Sized {
    // Required methods
    fn size_in_felts(&self) -> usize;
    fn encode_to(self, output: &mut Vec<StarkFelt>);
    fn decode(input: &mut FeltReader<'_>) -> Result<Self, FeltReaderError>;

    // Provided method
    fn into_encoded_vec(self) -> Vec<StarkFelt> { ... }
}

This codec allows to convert structured OS program output into array of felts

In order to prepare parameters for the Starknet contract updateState method: 1. Cast the output to dynamic uint256[] array 2. Get onchain data hash & size 3. ABI encode parameters

NOTE: Field element (252 bit) is encoded as an EVM word (256 bit) and vice versa EVM developer should be aware of that and prevent data loss by not using the higest 4 bits

Required Methods

fn size_in_felts(&self) -> usize

Return an estimation of the number of field elements required to encode self

This is to be used for allocating the correct amount of memory before encoding. It’s for optimization purpose (avoiding reallocation) so it’s implementation should be efficient (no iteration, no IO, no other allocation, no expensive

fn encode_to(self, output: &mut Vec<StarkFelt>)

Encodes current snos output field as felt array and appends to the result

fn decode(input: &mut FeltReader<'_>) -> Result<Self, FeltReaderError>

Tries to decode snos output field given a felt reader instance

Provided Methods

fn into_encoded_vec(self) -> Vec<StarkFelt>

Converts structured snos program output into array of field elements